requirements.txt
Flask
mysql-connector-python
schedule
